% Copyright 2001 by Roger S. Bivand
\name{columbus}
\alias{columbus}
\alias{col.gal.nb}
\alias{coords}
\docType{data}
\title{Columbus OH spatial analysis data set}
\usage{data(columbus)}
\description{
The data set is now part of the spData package
}
\examples{
example(columbus, package="spData")
}


\keyword{datasets}
