% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localC.R
\name{localC}
\alias{localC}
\alias{localC_perm}
\title{Compute Local Geary statistic}
\usage{
localC(x, listw, spChk = NULL)

localC_perm(x, listw, nsim = 499, spChk = NULL)
}
\arguments{
\item{x}{a numeric vector the same length as the neighbours list in listw}

\item{listw}{a \code{listw} object created for example by \code{nb2listw}}

\item{spChk}{hould the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use \code{get.spChkOption()}}
}
\description{
The Local Geary is a local adaptation of Geary's C statistic of spatial autocorrelation. The Local Geary uses squared differences to measure dissimilarity unlike the Local Moran. Low values of the Local Geary indicate positive spatial autocorrelation and large, negative.
}
\references{
{Anselin, L. (1995), Local Indicators of Spatial Association—LISA. Geographical Analysis, 27: 93-115. \doi{10.1111/j.1538-4632.1995.tb00338.x}}

{Anselin, L. (2019), A Local Indicator of Multivariate Spatial Association: Extending Geary's c. Geogr Anal, 51: 133-150. \doi{10.1111/gean.12164}}
}
\author{
Josiah Parry, \email{josiah.parry@gmail.com}
}
