\name{spdep}
\alias{spdep}
\title{Return package version number}
\description{
  The function retreives package version and build information
}
\usage{
spdep(build = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{build}{if TRUE, also returns build information}
}

\value{
  a character vector with one or two elements
}

\author{Roger Bivand \email{Roger.Bivand@nhh.no}}

\keyword{spatial}

