% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount-uni-impl.R
\name{local_joincount_uni}
\alias{local_joincount_uni}
\title{Calculate the local univariate join count}
\usage{
local_joincount_uni(
  x,
  listw,
  nsim = 199,
  alternative = "two.sided"
)
}
\arguments{
\item{x}{a binary variable either numeric or logical}

\item{listw}{a listw object containing binary weights created, for example, with \code{nbwlistw(nb, style = "B")}}

\item{nsim}{the number of conditional permutation simulations}

\item{alternative}{default \code{"greater"}. One of \code{"less"} or \code{"greater"}.}
}
\value{
a \code{data.frame} with two columns \code{join_count} and \code{p_sim} and number of rows equal to the length of \code{x}.
}
\description{
The univariate local join count statistic is used to identify clusters of rarely occurring binary variables. The binary variable of interest should occur less
than half of the time.
}
\details{
The local join count statistic requires a binary weights list which can be generated with \code{nb2listw(nb, style = "B")}. Additionally, ensure that the binary variable of interest is rarely occurring in no more than half of observations.

P-values are estimated using a conditional permutation approach. This creates a reference distribution from which the observed statistic is compared. For more see \href{https://geodacenter.github.io/glossary.html#ppvalue}{Geoda Glossary}.
}
\examples{
data(oldcol)
x <- ifelse(COL.OLD$CRIME < 35, 0L, 1L)
listw <- nb2listw(COL.nb, style = "B")
local_joincount_uni(x, listw)
}

\references{
Anselin, L., & Li, X. (2019). Operational Local Join Count Statistics for Cluster Detection. Journal of geographical systems, 21(2), 189–210. \doi{10.1007/s10109-019-00299-x}
}
\author{Josiah Parry \email{josiah.parry@gmail.com}}
