\name{spdep-defunct}
\alias{spdep-defunct}
\alias{lextrB}
\alias{lextrW}
\alias{lextrS}
\alias{l_max}
\alias{griffith_sone}
\alias{subgraph_eigenw}
\alias{mom_calc}
\alias{mom_calc_int2}
\alias{stsls}
\alias{print.stsls}
\alias{print.summary.stsls}
\alias{summary.stsls}
\alias{residuals.stsls}
\alias{coef.stsls}
\alias{deviance.stsls}
\alias{impacts.stsls}
\alias{GMerrorsar}
\alias{residuals.gmsar}
\alias{deviance.gmsar}
\alias{coef.gmsar}
\alias{fitted.gmsar}
\alias{print.gmsar}
\alias{summary.gmsar}
\alias{print.summary.gmsar}
\alias{Hausman.test.gmsar}
\alias{impacts.gmsar}
\alias{GMargminImage}
\alias{gstsls}
\alias{impacts.lagmess}
\alias{lagmess}
\alias{print.lagmess}
\alias{print.summary.lagmess}
\alias{summary.lagmess}
\alias{residuals.lagmess}
\alias{deviance.lagmess}
\alias{coef.lagmess}
\alias{fitted.lagmess}
\alias{logLik.lagmess}
\alias{ME}
\alias{print.ME_res}
\alias{fitted.ME_res}
\alias{SpatialFiltering}
\alias{print.SFResult}
\alias{fitted.SFResult}
\alias{LR.sarlm}
\alias{LR1.sarlm}
\alias{Wald1.sarlm}
\alias{Hausman.test}
\alias{Hausman.test.sarlm}
\alias{logLik.sarlm}
\alias{as.spam.listw}
\alias{as_dgRMatrix_listw}
\alias{as_dsTMatrix_listw}
\alias{as_dsCMatrix_I}
\alias{as_dsCMatrix_IrW}
\alias{Jacobian_W}
\alias{coerce,listw,CsparseMatrix-method}
\alias{coerce,listw,RsparseMatrix-method}
\alias{coerce,listw,symmetricMatrix-method}
\alias{powerWeights}
\alias{impacts.lagmess}
\alias{plot.lagImpact}
\alias{print.lagImpact}
\alias{summary.lagImpact}
\alias{print.summary.lagImpact}
\alias{HPDinterval.lagImpact}
\alias{can.be.simmed}
\alias{eigenw}
\alias{similar.listw}
\alias{do_ldet}
\alias{jacobianSetup}
\alias{eigen_setup}
\alias{eigen_pre_setup}
\alias{mcdet_setup}
\alias{cheb_setup}
\alias{spam_setup}
\alias{spam_update_setup}
\alias{Matrix_setup}
\alias{Matrix_J_setup}
\alias{LU_setup}
\alias{LU_prepermutate_setup}
\alias{moments_setup}
\alias{SE_classic_setup}
\alias{SE_whichMin_setup}
\alias{SE_interp_setup}
\alias{MCMCsamp}
\alias{MCMCsamp.spautolm}
\alias{MCMCsamp.sarlm}
\alias{spautolm}
\alias{residuals.spautolm}
\alias{deviance.spautolm}
\alias{coef.spautolm}
\alias{fitted.spautolm}
\alias{print.spautolm}
\alias{summary.spautolm}
\alias{LR1.spautolm}
\alias{logLik.spautolm}
\alias{print.summary.spautolm}
\alias{spBreg_sac}
\alias{impacts.MCMC_sar_g}
\alias{impacts.MCMC_sem_g}
\alias{impacts.MCMC_sac_g}
\alias{spBreg_err}
\alias{spBreg_lag}
\alias{predict.SLX}
\alias{lmSLX}
\alias{impacts.SLX}
\alias{create_WX}
\alias{anova.sarlm}
\alias{bptest.sarlm}
\alias{errorsarlm}
\alias{impacts}
\alias{impacts.sarlm}
\alias{intImpacts}
\alias{lagsarlm}
\alias{predict.sarlm}
\alias{print.sarlm.pred}
\alias{as.data.frame.sarlm.pred}
\alias{residuals.sarlm}
\alias{deviance.sarlm}
\alias{coef.sarlm}
\alias{vcov.sarlm}
\alias{fitted.sarlm}
\alias{sacsarlm}
\alias{summary.sarlm}
\alias{print.sarlm}
\alias{print.summary.sarlm}
\alias{trW}
\alias{aple.mc}
\alias{aple.plot}
\alias{localAple}
\alias{aple}


\title{Defunct Functions in Package \pkg{spdep}}

\description{
  These functions are defunct from release 1.2-1. The
  functions have been moved to the \pkg{spatialreg} package.
}

\usage{
aple.plot()
localAple()
aple.mc()
aple()
lextrB()
lextrW()
lextrS()
griffith_sone()
subgraph_eigenw()
mom_calc()
mom_calc_int2()
stsls()
\method{impacts}{stsls}()
GMerrorsar()
\method{summary}{gmsar}()
gstsls() 
\method{impacts}{gmsar}()
\method{Hausman.test}{gmsar}()
lagmess()
ME()
SpatialFiltering()
LR.sarlm()
\method{logLik}{sarlm}()
LR1.sarlm()
Wald1.sarlm()
\method{Hausman.test}{sarlm}()
as.spam.listw()
as_dgRMatrix_listw()
as_dsTMatrix_listw()
as_dsCMatrix_I()
as_dsCMatrix_IrW()
Jacobian_W()
powerWeights()
\method{plot}{lagImpact}()
\method{print}{lagImpact}()
\method{summary}{lagImpact}()
\method{HPDinterval}{lagImpact}()
intImpacts()
can.be.simmed()
eigenw()
similar.listw()
do_ldet()
jacobianSetup()
cheb_setup()
mcdet_setup()
eigen_setup()
eigen_pre_setup()
spam_setup()
spam_update_setup()
Matrix_setup()
Matrix_J_setup()
LU_setup()
LU_prepermutate_setup()
moments_setup()
SE_classic_setup()
SE_whichMin_setup()
SE_interp_setup()
MCMCsamp()
\method{MCMCsamp}{spautolm}()
\method{MCMCsamp}{sarlm}()
spautolm()
\method{summary}{spautolm}()
spBreg_sac()
\method{impacts}{MCMC_sar_g}()
\method{impacts}{MCMC_sem_g}()
\method{impacts}{MCMC_sac_g}()
spBreg_err()
spBreg_lag()
\method{predict}{SLX}()
lmSLX()
\method{impacts}{SLX}()
create_WX()
\method{anova}{sarlm}()
bptest.sarlm()
errorsarlm()
\method{impacts}{sarlm}()
lagsarlm()
\method{predict}{sarlm}()
\method{print}{sarlm.pred}()
\method{as.data.frame}{sarlm.pred}()
\method{residuals}{sarlm}()
\method{deviance}{sarlm}()
\method{coef}{sarlm}()
\method{vcov}{sarlm}()
\method{fitted}{sarlm}()
sacsarlm()
\method{summary}{sarlm}()
\method{print}{sarlm}()
\method{print}{summary.sarlm}()
trW()
}


\details{
 Model-fitting functions and functions supporting model fitting have been moved to the \pkg{spatialreg} package.
}


\seealso{
  \code{\link[base]{Defunct}}
}


\keyword{spatial}

